+++ src/Components/Playlist/Playlist.jsx
@@ -1,6 +1,6 @@
 import { useEffect, useRef, useState } from "react";
 import { Link } from "react-router-dom";
-import { useDataContext } from "../../Context/data-context";
+import { useDataContext } from "../../Context";
 import { PlaylistCard } from "./PlaylistCard";
 
 export const Playlist = () => {
