+++ src/Components/Video/FilterVideos.jsx
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import { useDataContext } from "../../Context/data-context";
+import { useDataContext } from "../../Context";
 
 export const FilterVideos = () => {
   const [searchTxt, setSearchTxt] = useState("");
